Rename Even or odd number.png to even or odd number.png
We first include the iostream header file to allow for input-output operations.
We declare an integer variable num to store the user's input.
We prompt the user to enter an integer using std::cout.
We use std:cin to read the user's input and store it in num.
We use the modulo operator (%) to check if the remainder of num divided by 2 is 0. If it is, the number is even.
We use an if statement to print out whether the number is even or odd.