Rename AND operation.png to And operation.png
The program prompts the user to enter a number and checks if the input number is between 10 and 20 (inclusive). If the number falls within this range, it displays a message indicating that the number is between 10 and 20. Otherwise, it displays a message stating that the number is not within this range.