Rename Positive or negative number.png to Positive or  negative number.png
This version of the program includes a space between "The number" and the number itself, which makes the output easier to read. It also correctly spells "positive". However, it still does not handle the case where the user enters a non-integer value. To handle this case, you could use a try-catch block or check the state of the input stream after attempting to read an integer.