Rename student information.png to Student information.png
Overview The program defines a Student class with properties (data members) for a student's name, roll number, age, and division. It also includes a displayInfo() function to print out the student's information. In the main() function, two instances of the Student class, s1 and s2, are created, and their properties are set. Finally, the program displays the information for each student using the displayInfo() function.