Rename student imformation.png to Student imformation.png
This C++ program defines a Student class with private data members: name, rollNo, and marks. It has two public member functions: inputDetails() to input the student's details and displayDetails() to display them. In the main() function, two Student objects (student1 and student2) are created. The program prompts the user to input details for both students and then displays their respective details. This implementation demonstrates basic object-oriented concepts like encapsulation and the use of classes and objects.