Rename syntex of switch.png to Syntex of switch.png
SHORT COMMENT ON THIS CODE

ChatGPT said:
ChatGPT
This C++ program uses a switch statement to print the name of a day based on the value of the day variable. It checks the value of day and matches it with the cases (1 through 7) to print the corresponding weekday name. If the value of day is 4, it will print "Thursday". If day has a value outside of 1-7, it will print "Invalid day". The break statement ensures that only the matched case is executed, and the program exits the switch block immediately afterward.